requirements.txt
# Corrected dependencies
streamlit==1.32.0
google-cloud-firestore==2.13.0
google-cloud-aiplatform>=1.68.0  # Vertex AI main package
python-dotenv==1.0.0
#google-generativeai>=0.3.0
#generativeai
